remove error badge
