+++ src/Aspire.Dashboard/Components/Pages/Resources.razor.css
@@ -1,8 +1,3 @@
-::deep .resource-state-container {
-    display: flex;
-    align-items: center;
-}
-
 ::deep .resource-state-badge {
     padding: 0 3px;
     cursor: pointer;
