+++ src/Aspire.Dashboard/Components/ResourcesGridColumns/StateColumnDisplay.razor
@@ -3,27 +3,25 @@
 @using Aspire.Dashboard.Resources
 @inject IStringLocalizer<Columns> Loc
 
-<div class="resource-state-container">
-    @if (Resource is { State: ResourceStates.ExitedState /* containers */ or ResourceStates.FinishedState /* executables */ })
-    {
-        if (Resource.TryGetExitCode(out int exitCode) && exitCode is not 0)
-        {
-            <!-- process completed unexpectedly, hence the non-zero code. this is almost certainly an error, so warn users -->
-            <FluentIcon Title="@string.Format(Loc[Columns.StateColumnResourceExitedUnexpectedly], Resource.ResourceType, exitCode)"
-                        Icon="Icons.Filled.Size16.ErrorCircle"
-                        Color="Color.Error"
-                        Class="severity-icon" />
-        }
-        else
-        {
-            <!-- process completed, which may not have been unexpected -->
-            <FluentIcon Title="@string.Format(Loc[Columns.StateColumnResourceExited], Resource.ResourceType)"
-                        Icon="Icons.Filled.Size16.Warning"
-                        Color="Color.Warning"
-                        Class="severity-icon" />
-        }
-    }
+@{
+    var unviewedErrorCount = GetUnviewedErrorCount(Resource);
+}
 
-    @Resource.State
-    <UnreadLogErrorsBadge UnviewedErrorCounts="UnviewedErrorCounts" Resource="@Resource" />
-</div>
+@if (unviewedErrorCount > 0)
+{
+    <FluentCounterBadge Count="@unviewedErrorCount" Appearance="Appearance.Lightweight" BackgroundColor="@Color.Error" Color="Color.Fill" Max="9">
+        <StateColumnText Resource="@Resource"/>
+    </FluentCounterBadge>
+}
+else
+{
+    <StateColumnText Resource="@Resource"/>
+}
+
+@code {
+    [Parameter, EditorRequired]
+    public required Dictionary<OtlpApplication, int>? UnviewedErrorCounts { get; set; }
+
+    [Parameter, EditorRequired]
+    public required ResourceViewModel Resource { get; set; }
+}
