+++ src/Aspire.Dashboard/Components/ResourcesGridColumns/StateColumnDisplay.razor.cs
@@ -2,16 +2,34 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Aspire.Dashboard.Model;
-using Aspire.Dashboard.Otlp.Model;
+using Aspire.Dashboard.Otlp.Storage;
 using Microsoft.AspNetCore.Components;
 
 namespace Aspire.Dashboard.Components.ResourcesGridColumns;
 
 public partial class StateColumnDisplay
 {
-    [Parameter, EditorRequired]
-    public required Dictionary<OtlpApplication, int>? UnviewedErrorCounts { get; set; }
+    [Inject]
+    public required TelemetryRepository TelemetryRepository { get; init; }
 
-    [Parameter, EditorRequired]
-    public required ResourceViewModel Resource { get; set; }
+    private int GetUnviewedErrorCount(ResourceViewModel resource)
+    {
+        if (UnviewedErrorCounts is null)
+        {
+            return 0;
+        }
+
+        var application = TelemetryRepository.GetApplication(resource.Uid);
+        if (application is null)
+        {
+            return 0;
+        }
+
+        if (!UnviewedErrorCounts.TryGetValue(application, out var count))
+        {
+            return 0;
+        }
+
+        return count;
+    }
 }
