--- src/Aspire.Dashboard/Components/ResourcesGridColumns/UnreadLogErrorsBadge.razor
@@ -1,19 +1,0 @@
-﻿@using Aspire.Dashboard.Resources
-@using Aspire.Dashboard.Model
-@namespace Aspire.Dashboard.Components
-@inject IStringLocalizer<Columns> Loc
-
-@if (_unviewedCount > 0)
-{
-    <FluentAnchor aria-label="@string.Format(Loc[_unviewedCount == 1 ? nameof(Columns.UnreadLogErrorsBadgeErrorLog) : nameof(Columns.UnreadLogErrorsBadgeErrorLogs)], _unviewedCount)"
-                  Href="@GetResourceErrorStructuredLogsUrl(Resource)"
-                  Appearance="Appearance.Stealth"
-                  Class="unread-logs-errors-link">
-        <FluentBadge
-            Appearance="Appearance.Accent"
-            Circular="true"
-            Fill="error">
-            @(_unviewedCount > 9 ? "9+" : _unviewedCount.ToString())
-        </FluentBadge>
-    </FluentAnchor>
-}
